+++ app/Models/Language.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Language extends Model
+{
+    /**
+     * Get the offers the model belongs to.
+     */
+    public function offers()
+    {
+        return $this->belongsToMany('App\Models\Offer');
+    }
+}