+++ database/migrations/2021_03_03_141705_redesign_languages_table.php
@@ -1,0 +1,41 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class RedesignLanguagesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('languages', function (Blueprint $table) {
+            $table->id();
+            $table->string('identifier')->unique();
+            $table->timestamps();
+        });
+
+        Schema::table('offers', function (Blueprint $table) {
+            $table->unsignedBigInteger('language_id');
+            $table->foreign('language_id')->references('id')->on('languages');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('offers', function (Blueprint $table) {
+            $table->dropForeign('offers_language_id_foreign');
+            $table->dropColumn('language_id');
+        });
+        Schema::dropIfExists('languages');
+    }
+}