+++ database/migrations/2021_03_03_150030_redesign_timestamps_table.php
@@ -1,0 +1,40 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class RedesignTimestampsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('timestamps', function (Blueprint $table) {
+            $table->id();
+            $table->unsignedBigInteger('offer_id');
+            $table->foreign('offer_id')
+                ->references('id')->on('offers')->unique();
+            $table->timestamp('executed_from')->default(DB::raw('CURRENT_TIMESTAMP'));
+            $table->timestamp('executed_until')->nullable();
+            $table->timestamp('listed_from')->default(DB::raw('CURRENT_TIMESTAMP'));
+            $table->timestamp('listed_until')->nullable();
+            $table->boolean('active')->default(true);
+            $table->timestamps();
+        });
+
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('timestamps');
+    }
+}