+++ database/migrations/2021_03_03_151454_redesign_hub_offer_table.php
@@ -1,0 +1,38 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class RedesignHubOfferTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('huboffers', function (Blueprint $table) {
+            $table->id();
+            $table->unsignedBigInteger('offer_id');
+            $table->foreign('offer_id')
+                ->references('id')->on('offers')->unique();
+            $table->integer('sort_flag')->nullable();
+            $table->text('keywords')->nullable();
+            $table->boolean('visible')->default(true);
+            $table->timestamps();
+        });
+
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('huboffers');
+    }
+}